fix: Explicitly cast steps_back_local to int to prevent slice TypeError

+++ tta.py
@@ -29,7 +29,7 @@ def tta_reconstruct(x, lion, steps_back_local, gamma, eta, p, total=100):
     )
     scheduler.set_timesteps(total, device='cuda')
     
-    steps_back_local = (total * steps_back_local) // 100
+    steps_back_local = int((total * steps_back_local) // 100)
     timesteps_local = scheduler.timesteps[-steps_back_local:]
     alpha_bar_local = scheduler.alphas_cumprod[timesteps_local[0]]
